Refactor external file importing and add API for desktop app
